refactor(home): change height detail article, refactor flow login and logout
