+++ app/src/main/AndroidManifest.xml
@@ -51,7 +51,7 @@
             android:exported="false" />
 
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.MainActivity"
             android:exported="false" />
     </application>
