+++ app/src/main/java/com/example/androidgroup4/ui/auth/LoginActivity.kt
@@ -8,6 +8,7 @@ import androidx.viewbinding.ViewBinding
 import com.example.androidgroup4.R
 import com.example.androidgroup4.base.BaseActivity
 import com.example.androidgroup4.databinding.ActivityLoginBinding
+import com.example.androidgroup4.ui.MainActivity
 import com.example.androidgroup4.utils.constant.PreferenceKeys.IS_LOGIN
 import com.example.androidgroup4.utils.getAppPreferenceEditor
 import com.example.androidgroup4.utils.isFormValid
@@ -47,6 +48,7 @@ class LoginActivity : BaseActivity<ActivityLoginBinding>() {
 
                 isFormValid(listOf(tilEmail, tilPassword)) {
                     getAppPreferenceEditor(this@LoginActivity).putBoolean(IS_LOGIN, true).commit()
+                    MainActivity.start(this@LoginActivity)
                     finish()
                 }
             }
