+++ app/src/main/java/com/example/androidgroup4/ui/onboarding/OnBoardingActivity.kt
@@ -5,12 +5,11 @@ import android.content.Intent
 import android.view.LayoutInflater
 import androidx.viewbinding.ViewBinding
 import androidx.viewpager2.widget.ViewPager2
-import com.example.androidgroup4.MainActivity
+import com.example.androidgroup4.ui.MainActivity
 import com.example.androidgroup4.R
 import com.example.androidgroup4.base.BaseActivity
 import com.example.androidgroup4.databinding.ActivityOnBoardingBinding
 import com.example.androidgroup4.ui.adapter.PagerAdapter
-import com.example.androidgroup4.ui.auth.LoginActivity
 
 class OnBoardingActivity : BaseActivity<ActivityOnBoardingBinding>() {
